requirements.txt
python-dotenv
psycopg2-binary
sqlalchemy
pandas
numpy
kagglehub
yfinance
sqlglot
langchain-community
langchain-experimental
langgraph